feat: add building and terrain components for world view
- Introduced BuildingActor component for rendering buildings with dynamic roofs and materials.
- Added FocusReticle component to display the currently followed agent.
- Implemented MinimapOverlay for a visual representation of the world with buildings and agents.
- Created ScreenSpaceCamera for managing camera positioning in the world.
- Developed TerrainLayer to render terrain tiles with water effects.
- Established WorldScene to manage the overall scene, including buildings and agents.
- Added WorldText component for displaying text in the 3D world.
- Implemented useTerrain hook for terrain generation based on building positions.
- Created useWorldSprites hook for managing agent sprites in the world.
- Defined styles for various building types in styles.ts.
- Added types for terrain tiles, building styles, and world scene props in types.ts.
- Implemented utility functions for coordinate transformations and geometry creation in utils.ts.
