+++ claudeville/css/layout.css
@@ -21,6 +21,7 @@ body {
     flex: 1;
     display: flex;
     flex-direction: column;
+    position: relative;
     min-width: 0;
     min-height: 0;
     overflow: hidden;
