+++ claudeville/src/presentation/react/world/components/FocusReticle.tsx
@@ -1,0 +1,5 @@
+export function FocusReticle({ label }: { label: string }) {
+  return (
+    <div className="world-view__focus-badge">Following {label}</div>
+  );
+}
